TeamG2_DialogAnalyser-main/requirements.txt
protobuf~=5.29.6
python-dotenv~=1.2.1
google-generativeai~=0.8.6